refactor: modularize hybrid bot
